backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pandas==2.2.0
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
google-generativeai==0.3.2
yagmail==0.15.293
sqlalchemy==2.0.25
openpyxl==3.1.2
tenacity==8.2.3
python-dotenv==1.0.1
